Extract TempSet from llcallbacklist.cpp into its own tempset.h.
